requirements.txt
anyio==3.5.0
asgiref==3.5.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
djangorestframework==3.13.1
fastapi==0.74.1
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
platformdirs==2.5.1
psycopg2==2.9.3
pydantic==1.9.0
pytz==2022.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
typing_extensions==4.1.1
tzdata==2022.1
virtualenv==20.13.1
